the cleanup never ends
