+++ Assets/Scripts/Pickups/HealthPotion.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public class HealthPotion : Pickup
