+++ Assets/Scripts/Pickups/ManaPotion.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public class ManaPotion : Pickup
