+++ Assets/Scripts/Pickups/SpeedPotion.cs
@@ -1,5 +1,4 @@
 using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public class SpeedPotion : Pickup
