+++ Assets/Scripts/Signals.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b614f4d66cddd4a76b29abb3391a8c01
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
