+++ Assets/Scripts/Signals/DialogueSignalSO.cs
@@ -1,0 +1,5 @@
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(menuName = "Signals/Dialogue Signal")]
+public class DialogueSignalSO : SignalSO<List<string>> {}
