+++ Assets/Scripts/Signals/DialogueSignalSO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cd3b45afeddd04958acfea26da7dca1c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
