+++ Assets/Scripts/Signals/LoadSceneSignalSO.cs
@@ -1,0 +1,4 @@
+using UnityEngine;
+
+[CreateAssetMenu(menuName = "Signals/Load Scene Signal")]
+public class LoadSceneSignalSO : SignalSO<SceneSO> {}
