+++ Assets/Scripts/Signals/LoadSceneSignalSO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 67480fdd26e294ab1b222c8dc2d472ae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
