+++ Assets/Scripts/Signals/SpawnPointSignalSO.cs
@@ -1,0 +1,4 @@
+using UnityEngine;
+
+[CreateAssetMenu(menuName = "Signals/Spawn Point Signal")]
+public class SpawnPointSignal : SignalSO<Vector3> {}
