+++ Assets/Scripts/Signals/SpawnPointSignalSO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7c2a06803e16847efb6e476530975a94
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
